chore: Split main.rs
This is the first contribution in a refactoring series to improve the
modularity and error handling of the crate. This moves the RecordState
into the newly introduces `lib.rs`.
